% Auto-generated: do not edit by hand
\name{allotaxonometerForAll}

\alias{allotaxonometerForAll}

\title{AllotaxonometerForAll component}

\description{

}

\usage{
allotaxonometerForAll(id=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Unnamed list. The allotaxonometer data. Should have the form:

  `[{"x1":0,"y1":0,"coord_on_diag":0,"cos_dist":0,"rank":"(1, 1)","rank_L":[1,1],"rank_R":[1,1],"value":1,"types":"Mary","which_sys":"right"}...]`}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

